+++ packages/cwrc-leafwriter/src/components/sourceEditor/XmlMonacoEditor.tsx
@@ -418,7 +418,9 @@ export const XmlMonacoEditor = ({
 
   useEffect(() => {
     editor?.updateOptions({
-      fontFamily: tibetanFont ? TIBETAN_EDITOR_FONT_FAMILY : 'Menlo, Monaco, "Courier New", monospace',
+      fontFamily: tibetanFont
+        ? TIBETAN_EDITOR_FONT_FAMILY
+        : 'Menlo, Monaco, "Courier New", monospace',
     });
   }, [editor, tibetanFont]);